+++ src/Arborescence.Primitives/IndexedSet.cs
@@ -1,57 +1,51 @@
-namespace Arborescence.Models
+namespace Arborescence
 {
     using System;
     using System.Collections;
     using System.Collections.Generic;
-    using System.Runtime.CompilerServices;
 
     /// <summary>
-    /// Represents a set of values as a bit array.
+    /// Represents a set of values as a byte array.
     /// </summary>
 #if NET5
-    public readonly struct CompactSet : IReadOnlySet<int>, ISet<int>, IEquatable<CompactSet>
+    public readonly struct IndexedSet : IReadOnlySet<int>, ISet<int>, IEquatable<IndexedSet>
 #else
-    public readonly struct CompactSet : ISet<int>, IEquatable<CompactSet>
+    public readonly struct IndexedSet : ISet<int>, IEquatable<IndexedSet>
 #endif
     {
-        private const int BitShiftPerByte = 3;
-
         private readonly byte[] _items;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="CompactSet"/> structure.
+        /// Initializes a new instance of the <see cref="IndexedSet"/> structure.
         /// </summary>
         /// <param name="items">The backing store for the set.</param>
-        public CompactSet(byte[] items) => _items = items;
-
-        /// <summary>
-        /// Get the number of bytes required to hold <paramref name="count"/> bit values.
-        /// </summary>
-        /// <param name="count">The number of bit values.</param>
-        /// <returns>The number of bytes.</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int GetByteCount(int count)
+        /// <exception cref="ArgumentNullException"><paramref name="items"/> is <see langword="null"/>.</exception>
+        public IndexedSet(byte[] items)
         {
-            if (count <= 0)
-                return 0;
+            if (items is null)
+                ThrowHelper.ThrowArgumentNullException(ExceptionArgument.items);
 
-            uint temp = (uint)(count - 1 + (1 << BitShiftPerByte));
-            return (int)(temp >> BitShiftPerByte);
+            _items = items;
         }
 
         /// <inheritdoc/>
-        public IEnumerator<int> GetEnumerator() => throw new NotSupportedException();
+        public IEnumerator<int> GetEnumerator()
+        {
+            for (int i = 0; i < _items.Length; ++i)
+            {
+                if (_items[i] != 0)
+                    yield return i;
+            }
+        }
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
         void ICollection<int>.Add(int item)
         {
-            int byteIndex = Div8Rem(item, out int bitIndex);
-            if (unchecked((uint)byteIndex >= (uint)_items.Length))
-                return;
+            if ((uint)item >= (uint)_items.Length)
+                throw new ArgumentOutOfRangeException(nameof(item));
 
-            byte bitMask = (byte)(1u << bitIndex);
-            _items[byteIndex] |= bitMask;
+            _items[item] = 1;
         }
 
         /// <inheritdoc/>
@@ -84,46 +78,45 @@ namespace Arborescence.Models
         /// <inheritdoc/>
         public void UnionWith(IEnumerable<int> other) => throw new NotSupportedException();
 
-        bool ISet<int>.Add(int item)
-        {
-            int byteIndex = Div8Rem(item, out int bitIndex);
-            if (unchecked((uint)byteIndex >= (uint)_items.Length))
-                return false;
+        bool ISet<int>.Add(int item) => Add(item);
 
-            byte bitMask = (byte)(1u << bitIndex);
-            bool result = (_items[byteIndex] & bitMask) == 0;
-            _items[byteIndex] |= bitMask;
+        /// <inheritdoc cref="ISet{T}"/>
+        public bool Add(int item)
+        {
+            if ((uint)item >= (uint)_items.Length)
+                throw new ArgumentOutOfRangeException(nameof(item));
+
+            bool result = _items[item] == 0;
+            _items[item] = 1;
             return result;
         }
 
         /// <summary>
-        /// Removes all elements from a <see cref="CompactSet"/> object.
+        /// Removes all elements from a <see cref="IndexedSet"/> object.
         /// </summary>
         public void Clear() => Array.Clear(_items, 0, _items.Length);
 
         /// <summary>
-        /// Determines whether a <see cref="CompactSet"/> object contains the specified element.
+        /// Determines whether a <see cref="IndexedSet"/> object contains the specified element.
         /// </summary>
-        /// <param name="item">The element to locate in the <see cref="CompactSet"/> object.</param>
+        /// <param name="item">The element to locate in the <see cref="IndexedSet"/> object.</param>
         /// <returns>
-        /// <see langword="true"/> if the <see cref="CompactSet"/> object contains the specified element;
+        /// <see langword="true"/> if the <see cref="IndexedSet"/> object contains the specified element;
         /// otherwise, <see langword="false"/>.
         /// </returns>
         public bool Contains(int item)
         {
-            int byteIndex = Div8Rem(item, out int bitIndex);
-            if (unchecked((uint)byteIndex >= (uint)_items.Length))
+            if (unchecked((uint)item >= (uint)_items.Length))
                 return false;
 
-            byte bitMask = (byte)(1u << bitIndex);
-            return (_items[byteIndex] & bitMask) != 0;
+            return _items[item] != 0;
         }
 
         /// <inheritdoc/>
         public void CopyTo(int[] array, int arrayIndex) => throw new NotSupportedException();
 
         /// <summary>
-        /// Removes the specified element from a <see cref="CompactSet"/> object.
+        /// Removes the specified element from a <see cref="IndexedSet"/> object.
         /// </summary>
         /// <param name="item">The element to remove.</param>
         /// <returns>
@@ -131,12 +124,10 @@ namespace Arborescence.Models
         /// </returns>
         public bool Remove(int item)
         {
-            int byteIndex = Div8Rem(item, out int bitIndex);
-            if (unchecked((uint)byteIndex >= (uint)_items.Length))
+            if (unchecked((uint)item >= (uint)_items.Length))
                 return false;
 
-            byte bitMask = unchecked((byte)~(1u << bitIndex));
-            _items[byteIndex] &= bitMask;
+            _items[item] = 0;
             return true;
         }
 
@@ -147,10 +138,10 @@ namespace Arborescence.Models
         public bool IsReadOnly => false;
 
         /// <inheritdoc/>
-        public bool Equals(CompactSet other) => Equals(_items, other._items);
+        public bool Equals(IndexedSet other) => Equals(_items, other._items);
 
         /// <inheritdoc/>
-        public override bool Equals(object obj) => obj is CompactSet other && Equals(other);
+        public override bool Equals(object obj) => obj is IndexedSet other && Equals(other);
 
         /// <inheritdoc/>
         public override int GetHashCode() => _items != null ? _items.GetHashCode() : 0;
@@ -164,7 +155,7 @@ namespace Arborescence.Models
         /// <see langword="true"/> if the underlying arrays are reference equal;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public static bool operator ==(CompactSet left, CompactSet right) => left.Equals(right);
+        public static bool operator ==(IndexedSet left, IndexedSet right) => left.Equals(right);
 
         /// <summary>
         /// Checks inequality between two instances.
@@ -175,14 +166,6 @@ namespace Arborescence.Models
         /// <see langword="true"/> if the underlying arrays are not reference equal;
         /// <see langword="false"/> otherwise.
         /// </returns>
-        public static bool operator !=(CompactSet left, CompactSet right) => !left.Equals(right);
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static int Div8Rem(int number, out int remainder)
-        {
-            uint quotient = (uint)number >> 3;
-            remainder = number & 0b111;
-            return (int)quotient;
-        }
+        public static bool operator !=(IndexedSet left, IndexedSet right) => !left.Equals(right);
     }
 }