+++ tests/Arborescence.Traversal.Tests/QueueGenericSearchEnumerateVerticesTest.cs
@@ -6,7 +6,6 @@ namespace Arborescence
     using System.Collections.Generic;
     using System.Diagnostics;
     using Misnomer;
-    using Models;
     using Traversal;
     using Xunit;
     using EdgeEnumerator = System.ArraySegment<Endpoints>.Enumerator;